RAG/requirements.txt
langchain
openai
faiss-cpu
sentence-transformers
beautifulsoup4
requests
python-dotenv
